pyproject.toml
[build-system]
requires = [ "poetry-core",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "adnexus"
version = "0.1.0"
description = ""
authors = [ "nl <nl@x-net.at>",]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.6.4"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
manage-version = "scripts.manage_version:manage_version"
test = "scripts.test:test_only"
coverage = "scripts.test:with_coverage"

[tool.poetry.group.tests.dependencies]
pytest = "^8.1.1"
coverage = "^7.4.4"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.23.6"
pytest-html = "^4.1.1"
starlette = "^0.37.2"

[tool.poetry.group.dev.dependencies]
toml = "^0.10.2"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"
mkdocstrings = {extras = ["python"], version = "^0.24.1"}
mkdocs-material = "^9.5.14"


[tool.poetry.group.lint.dependencies]
ruff = "^0.3.5"

[tool.coverage.run]
source = ["xdi"]
branch = false
omit = ["*/integrations/*", "*/tests/*"]

